==> notebooks/try_1/01_eda.ipynb <==
import pandas as pd

path = r'D:\0_Respaldo\0_Proyectos_2024\ML_proyects\KagelX\KaggelX_Challenge\data\train.csv'

data = pd.read_csv(path)

df = data.copy()

df.info()
df.describe()

df.head(3)

### EDA

df['accident'].value_counts()

df['clean_title'].value_counts()

df['engine'].value_counts()

df['transmission'].value_counts()

df['ext_col'].value_counts()

df['int_col'].value_counts()

df['accident'].value_counts()	

df['fuel_type'].value_counts()	

df.head(2)

del df['model']
del df['clean_title']

del df['id']

df.head(3)

path = r'D:\0_Respaldo\0_Proyectos_2024\ML_proyects\KagelX\KaggelX_Challenge\data\transform_data\clean_data.parquet'

df.to_parquet(path)
