==> notebooks/try_1/02_transform_categoric.ipynb <==
import pandas as pd
from sklearn.preprocessing import LabelEncoder

path = r'D:\0_Respaldo\0_Proyectos_2024\ML_proyects\KagelX\KaggelX_Challenge\data\transform_data\clean_data.parquet'

data = pd.read_parquet(path)

df =data.copy()

df.head(3)

df[df['brand'] == 'Mercedes-Benz'].value_counts()

### Transformation

label_brand = LabelEncoder()
label_fuel = LabelEncoder()
label_engine = LabelEncoder()
label_trans = LabelEncoder()
label_ext_col = LabelEncoder()
label_ext_int = LabelEncoder()
label_accident = LabelEncoder()


df['brand_t'] = label_brand.fit_transform(df['brand'])
df['fuel_type_t'] = label_fuel.fit_transform(df['fuel_type'])
df['engine_t'] = label_engine.fit_transform(df['engine'])
df['transmission_t'] = label_trans.fit_transform(df['transmission'])
df['ext_col_t'] = label_ext_col.fit_transform(df['ext_col'])
df['int_col_t'] = label_ext_int.fit_transform(df['int_col'])
df['accident_t'] = label_accident.fit_transform(df['accident'])

#Sanity check
print('Clases aprendidas:', label_brand.classes_)
print('Clases aprendidas:', label_engine.classes_)

df1 = df.copy()

del df1['brand']
del df1['engine']
del df1['transmission']
del df1['ext_col']
del df1['int_col']
del df1['accident']
del df1['fuel_type']


df1.head(5)

#Agregamos la columna +5000
df1['no_5k_millas'] = df['milage'].apply(lambda x: 1 if x < 5000 else 0)

df1.head(5)

path = r'D:\0_Respaldo\0_Proyectos_2024\ML_proyects\KagelX\KaggelX_Challenge\data\transform_data\transform_data.parquet'

df1.to_parquet(path)

### Label encoder export
Exporto el label encoder para utilizarlo despues en el modelo.

import joblib as jb

path = r'D:\0_Respaldo\0_Proyectos_2024\ML_proyects\KagelX\KaggelX_Challenge\data\transform_data\label_engine.pkl'
jb.dump(label_engine, path)

path = r'D:\0_Respaldo\0_Proyectos_2024\ML_proyects\KagelX\KaggelX_Challenge\data\transform_data\label_ext_int.pkl'
jb.dump(label_ext_int, path)

path = r'D:\0_Respaldo\0_Proyectos_2024\ML_proyects\KagelX\KaggelX_Challenge\data\transform_data\label_accident.pkl'
jb.dump(label_accident, path)

path = r'D:\0_Respaldo\0_Proyectos_2024\ML_proyects\KagelX\KaggelX_Challenge\data\transform_data\label_brand.pkl'
jb.dump(label_brand, path)

path = r'D:\0_Respaldo\0_Proyectos_2024\ML_proyects\KagelX\KaggelX_Challenge\data\transform_data\label_ext_col.pkl'
jb.dump(label_ext_col, path)

path = r'D:\0_Respaldo\0_Proyectos_2024\ML_proyects\KagelX\KaggelX_Challenge\data\transform_data\label_fuel.pkl'
jb.dump(label_fuel, path)

path = r'D:\0_Respaldo\0_Proyectos_2024\ML_proyects\KagelX\KaggelX_Challenge\data\transform_data\label_trans.pkl'
jb.dump(label_trans, path)
