==> notebooks/try_1/02_2_transform_correlaton.ipynb <==
import pandas as pd
from sklearn.preprocessing import LabelEncoder
import joblib as jb

path = r'D:\0_Respaldo\0_Proyectos_2024\ML_proyects\KagelX\KaggelX_Challenge\data\transform_data\clean_data.parquet'

data = pd.read_parquet(path)

data.head(3)

#Result of the correlation analysis the columns will be modelyear , milage, price, engine,accidents.

df = data[['model_year','milage','engine','accident','price']].copy()

df.head(3)

### Transform categorical

label_encoder = LabelEncoder()

#Categorical transform
df['engine_t'] = label_encoder.fit_transform(df['engine'])
df['accident_t'] = label_encoder.fit_transform(df['accident'])

# delete columns

del df['engine']

del df['accident']


df.head(5
        )

### Engeniering

#### Miles by year

df_price = df.pop('price')

df1 = df.copy()

# Milage by year col
df1['milage_by_year'] = round((df['milage']/(2025 - df['model_year'])),2)


df1.head(5)

#### -5000 miles and -35miles coulumns

df1['less_5k_millas'] = df['milage'].apply(lambda x: 1 if x < 5000 else 0)

df1['less_35k_millas'] = df['milage'].apply(lambda x: 1 if x < 35000 else 0)

df1.head(5)

df1.describe()

#### Concatenar files

df_f = pd.concat([df1,df_price], axis=1)


df_f

path = r'D:\0_Respaldo\0_Proyectos_2024\ML_proyects\KagelX\KaggelX_Challenge\data\transform_data\t2_data.parquet'

df_f.to_parquet(path)
