==> notebooks/try_1/03_ml.ipynb <==
import pandas as pd
from sklearn.preprocessing import LabelEncoder
import joblib as jb
import matplotlib.pyplot as plt

import numpy as np

from sklearn.model_selection import train_test_split # for splitting the data
from sklearn.metrics import mean_squared_error, mean_absolute_error # for calculating the cost function
from sklearn.tree import DecisionTreeRegressor # for building the model

path = r'D:\0_Respaldo\0_Proyectos_2024\ML_proyects\KagelX\KaggelX_Challenge\data\transform_data\transform_data.parquet'

data = pd.read_parquet(path)

df = data.copy()

df.head(3)

x = df.drop(['price'], axis=1) #Features
y = df['price'] #target

# Splitting the dataset into training and testing set (80/20)
x_train, x_test, y_train, y_test = train_test_split(x, y, test_size = 0.2, random_state = 28)

# Initializing the Decision Tree Regression model
model = DecisionTreeRegressor(criterion='poisson',random_state = 0, max_depth=2)

# Fitting the Decision Tree Regression model to the data
model.fit(x_train, y_train)

# Max_depth optimizer
from sklearn.model_selection import GridSearchCV
param_grid = {'max_depth': [2, 5, 6, 7, 8, 20]}   # Values to try 

grid_search = GridSearchCV(estimator=model, param_grid=param_grid, cv=5, scoring='neg_mean_squared_error')

grid_search.fit(x_train, y_train)

#find the best value for 'max_depth'
best_max_depth = grid_search.best_params_['max_depth']
print(" max_depth best value is:", best_max_depth)

# Predicting the target values of the test set
y_pred = model.predict(x_test)

# RMSE (Root Mean Square Error)
rmse = float(format(np.sqrt(mean_squared_error(y_test, y_pred)), '.3f'))
print("\nRMSE: ", rmse)

mae = mean_absolute_error(y_test, y_pred)
print("\nMAE: ", mae)

### Dataframe correlacionado

path = r'D:\0_Respaldo\0_Proyectos_2024\ML_proyects\KagelX\KaggelX_Challenge\data\transform_data\corr_data.parquet'

data = pd.read_parquet(path)

df = data.copy()

x = df.drop(['price'], axis=1) #Features
y = df['price'] #target

x_train, x_test, y_train, y_test = train_test_split(x, y, test_size = 0.2, random_state = 28)

# Initializing the Decision Tree Regression model
model2 = DecisionTreeRegressor(criterion='poisson',random_state = 0, max_depth=2)

# Fitting the Decision Tree Regression model to the data
model2.fit(x_train, y_train)

# Max_depth optimizer

param_grid = {'max_depth': [2, 5, 6, 7, 8, 20]}   # Values to try 

grid_search = GridSearchCV(estimator=model2, param_grid=param_grid, cv=5, scoring='neg_mean_squared_error')

grid_search.fit(x_train, y_train)

#find the best value for 'max_depth'
best_max_depth = grid_search.best_params_['max_depth']
print(" max_depth best value is:", best_max_depth)

# Predicting the target values of the test set
y_pred = model2.predict(x_test)

# RMSE (Root Mean Square Error)
rmse = float(format(np.sqrt(mean_squared_error(y_test, y_pred)), '.3f'))
print("\nRMSE: ", rmse)

mae = mean_absolute_error(y_test, y_pred)
print("\nMAE: ", mae)

### Dataframe transformacion 2 
Datos correlacionados, 3 columnas adicionales 

path = r'D:\0_Respaldo\0_Proyectos_2024\ML_proyects\KagelX\KaggelX_Challenge\data\transform_data\t2_data.parquet'

data = pd.read_parquet(path)

df = data.copy()

x = df.drop(['price'], axis=1) #Features
y = df['price'] #target

x_train, x_test, y_train, y_test = train_test_split(x, y, test_size = 0.2, random_state = 28)

print("Valores infinitos en x_train:", np.isinf(x_train).sum())

# Initializing the Decision Tree Regression model
model3 = DecisionTreeRegressor(criterion='poisson',random_state = 0, max_depth=2)

# Fitting the Decision Tree Regression model to the data
model3.fit(x_train, y_train)

# Predicting the target values of the test set
y_pred = model3.predict(x_test)

# RMSE (Root Mean Square Error)
rmse = float(format(np.sqrt(mean_squared_error(y_test, y_pred)), '.3f'))
print("\nRMSE: ", rmse)

mae = mean_absolute_error(y_test, y_pred)
print("\nMAE: ", mae)
