==> notebooks/try_1/04_ml_lightgbm_model.ipynb <==
import pandas as pd
import numpy as np

import joblib as jb

import lightgbm as lgb

from sklearn.model_selection import train_test_split # for splitting the data
from sklearn.metrics import mean_squared_error, mean_absolute_error # for calculating the cost function

from sklearn.model_selection import GridSearchCV

path = r'D:\0_Respaldo\0_Proyectos_2024\ML_proyects\KagelX\KaggelX_Challenge\data\transform_data\transform_data.parquet'

data = pd.read_parquet(path)

df = data.copy()

x = df.drop(['price'], axis=1) #Features
y = df['price'] #target

# Splitting the dataset into training and testing set (80/20)
x_train, x_test, y_train, y_test = train_test_split(x, y, test_size = 0.2, random_state = 28)

model = lgb.LGBMRegressor(learning_rate=0.05, 
                          num_leaves=31, 
                          max_depth=10, 
                          min_data_in_leaf=100,
                          bagging_fraction= 0.8, 
                          bagging_freq= 5,
                          feature_fraction= 0.6,
                          lambda_l2= 0.5,
                          n_estimators= 100
                          )


model.fit(x_train, y_train)

#### Entrenar hiperparametros

param_grid = {
    'learning_rate': [0.01, 0.05, 0.1],
    'num_leaves': [31, 50, 100],
    'max_depth': [-1, 10, 20],
    'min_data_in_leaf': [20, 50, 100],
    'feature_fraction': [0.6, 0.8, 1.0],
    'bagging_fraction': [0.6, 0.8, 1.0],
    'bagging_freq': [0, 5, 10],
    'lambda_l1': [0, 0.1, 0.5],
    'lambda_l2': [0, 0.1, 0.5],
    'n_estimators': [100, 200, 500]
}

param_grid_1 = {
    'lambda_l1': [0, 0.1, 0.5],
    'lambda_l2': [0, 0.1, 0.5],
    'n_estimators': [100, 200, 500]
    }

grid_search = GridSearchCV(
    estimator=model, 
    param_grid=param_grid_1, 
    scoring='neg_mean_squared_error', 
    cv=3, 
    verbose=1, 
    n_jobs=-1
)

grid_search.fit(x_train, y_train)

print("Best hiperparameters", grid_search.best_params_)

# Predicting the target values of the test set
y_pred = model.predict(x_test)

# RMSE (Root Mean Square Error)
rmse = float(format(np.sqrt(mean_squared_error(y_test, y_pred)), '.3f'))
print("\nRMSE: ", rmse)

mae = mean_absolute_error(y_test, y_pred)
print("\nMAE: ", mae)

#### Export model

jb.dump(model, r'D:\0_Respaldo\0_Proyectos_2024\ML_proyects\KagelX\KaggelX_Challenge\data\transform_data\ml.pkl')
