==> notebooks/try_1/05_Correlations.ipynb <==
import pandas as pd

path = r'D:\0_Respaldo\0_Proyectos_2024\ML_proyects\KagelX\KaggelX_Challenge\data\transform_data\transform_data.parquet'

data = pd.read_parquet(path)

df = data.copy()

df.head(5)

cor_1 = df.corr()

corTarget = abs(cor_1['price'])
relevant_features = corTarget[corTarget > 0.1]
relevant_features

df1 = df[['model_year','milage','engine_t','accident_t', 'price']]

path = r'D:\0_Respaldo\0_Proyectos_2024\ML_proyects\KagelX\KaggelX_Challenge\data\transform_data\corr_data.parquet'

df1.to_parquet(path)
